Move codemods to separate package
