+++ packages/cli/src/presets/codemods/__tests__/codemods.test.tsx
@@ -20,7 +20,7 @@ const expectCodemodToHaveBeenRan = (name: string, runPath: string) => {
   expect(execAsync).toHaveBeenCalledWith(
     expect.stringMatching(
       new RegExp(
-        `.*--transform=.*node_modules\\/@compiled\\/react\\/dist\\/cjs\\/codemods\\/${name}\\/index.js ${regexPath}`
+        `.*--transform=.*node_modules\\/@compiled\\/codemods\\/dist\\/${name}\\/index.js ${regexPath}`
       )
     )
   );
