+++ packages/codemods/CHANGELOG.md
@@ -1,0 +1,3 @@
+# @compiled/codemods
+
+## 0.6.15
