+++ packages/codemods/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@compiled/codemods",
+  "version": "0.6.14",
+  "description": "A familiar and performant compile time CSS-in-JS library for React.",
+  "author": "Michael Dougall",
+  "license": "Apache-2.0",
+  "homepage": "https://compiledcssinjs.com",
+  "bugs": "https://github.com/atlassian-labs/compiled/issues/new?assignees=&labels=bug&template=bug_report.md",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/atlassian-labs/compiled.git",
+    "directory": "packages/codemods"
+  },
+  "main": "./dist/cjs/index.js",
+  "sideEffects": false,
+  "keywords": [
+    "compiled",
+    "css-in-js",
+    "styled-components",
+    "emotion-js",
+    "typescript"
+  ],
+  "dependencies": {},
+  "devDependencies": {}
+}
