+++ packages/codemods/src/README.md
@@ -1,0 +1,8 @@
+# codemods
+
+> Codemods for easy migration from [styled components](https://styled-components.com/) and [emotion](https://emotion.sh/docs/introduction).
+
+## Available codemods
+
+1. [styled-components-to-compiled](./styled-components-to-compiled)
+2. [emotion-to-compiled](./emotion-to-compiled)
