+++ packages/codemods/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig.options.json",
+  "compilerOptions": {
+    "rootDir": "src",
+    "outDir": "dist",
+    "module": "commonjs"
+  }
+}
